requirements_web.txt
flask
opencv-python
torch
torchvision
ultralytics
pillow
numpy